Changing Cell type: Step 14

+++ exafmm/2d/types.h
@@ -7,7 +7,7 @@
 
 namespace exafmm {
   // Basic type definitions
-  typedef float real_t;                                         //!< Floating point type is single precision
+  typedef double real_t;                                         //!< Floating point type is single precision
   typedef std::complex<real_t> complex_t;                       //!< Complex type
 
   //! Structure of bodies
